requirements.txt
asgiref==3.8.1
Django
django-ckeditor==6.7.1
django-ckeditor-5==0.2.15
django-js-asset==2.2.0
pillow==11.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.2
tzdata==2024.2
djangorestframework==3.15.1
django-cors-headers==4.3.1